refactor: extract TinyOS terminal app
